% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_international_crs.R
\name{access_international_crs}
\alias{access_international_crs}
\title{Access database of international CRS data from R}
\usage{
access_international_crs()
}
\value{
Returns a remote tibble data frame linked to the international CRS database table.
}
\description{
This function connects to a specified DuckDB database table to retrieve CRS (Creditor Reporting System) activity level data of 
Official Development Finance (ODA/OOF) from official donors (DAC, non-DAC, multilaterals) and private philantropy, starting from 1960 to the recent year. 
It returns a remote tibble linked to the crs database file, and allows for using dplyr syntax in database queries.
The database file is expected to be located in a synchronized SharePoint directory on the user's local computer.
Use DBI::dbDisconnect(con, shutdown=TRUE) to close connection to database.
Data source: https://stats.oecd.org/DownloadFiles.aspx?DatasetCode=CRS1
}
\examples{
?access_crs()

}
