% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_statsys.R
\name{access_statsys}
\alias{access_statsys}
\title{Access database of Statsys data from R}
\usage{
access_statsys()
}
\value{
Returns a remote tibble data frame linked to the Statsys database table.
}
\description{
This function connects to a specified DuckDB database table to retrieve Statsys data. This includes official development assistance (ODA), frame agreement level data, Other official flows(OOF), export credits and private flows.
The data is available from 1960 to the recent year.
The database file is expected to be located in a synchronized SharePoint directory on the user's local computer.
Use DBI::dbDisconnect(con, shutdown=TRUE) to close connection to database.
}
\examples{
?access_statsys()

}
