% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_basic.R
\name{add_cols_basic}
\alias{add_cols_basic}
\title{Add basic columns to dataframe}
\usage{
add_cols_basic(data)
}
\arguments{
\item{data}{Input dataframe of Norwegian development assistance}
}
\value{
Returns dataframe with additional columns:
\itemize{
  \item target_area_no: Norwegian translation of column \emph{Target area}.
  \item partner_group_visual_no: Visual grouping and Norwegian translation of column \emph{Group of agreement partner}.
  \item partner_group_visual: Visual grouping of column \emph{Group of agreement partner}.
  \item main_region_no: Norwegian translation of column \emph{Main region}.
  \item income_category_no: Norwegain translation of column \emph{Income category}.
}
}
\description{
Add basic columns to dataframe
}
\examples{
?add_cols_basic()
}
