% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_basic_clean.R
\name{add_cols_basic_clean}
\alias{add_cols_basic_clean}
\title{Add basic columns to data frame of Norwegian development assistance}
\usage{
add_cols_basic_clean(data)
}
\arguments{
\item{data}{Input dataframe of Norwegian development assistance in snake_case column names}
}
\value{
Returns dataframe with additional columns:
\itemize{
  \item target_area_no: Norwegian translation of column \emph{Target area}.
  \item partner_group_visual_no: Visual grouping and Norwegian translation of column \emph{Group of agreement partner}.
  \item partner_group_visual: Visual grouping of column \emph{Group of agreement partner}.
  \item main_region_no: Norwegian translation of column \emph{Main region}.
  \item income_category_no: Norwegain translation of column \emph{Income category}.
  \item earmarked: Earmarked aid (bilateral, earmarked to multilaterals, triangular co-operation)
  \item countryspecified: Country specific aid (income category not unspecified)
  \item geographically_specified: Geographically specific aid (main region not unspecified)
  \item earmarked_subsaharan: Earmarked aid to Sub-Saharan Africa
  \item gross_disbursed_nok: Gross disbursed amount in Norwegian kroner. Converts negative amounts extended to zero.
  \item disbursed: Disbursed amount in US dollars. Based on column disbursed_1000.
}
}
\description{
Add basic columns to data frame of Norwegian development assistance
}
\examples{
?add_cols_basic_clean()
}
