% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_climate.R
\name{add_cols_climate}
\alias{add_cols_climate}
\title{Add climate columns to dataframe}
\usage{
add_cols_climate(data)
}
\arguments{
\item{data}{Input dataframe of Norwegian development assistance}
}
\value{
Returns dataframe with additional climate-relevant columns:
\itemize{
  \item climate_aid_type: Type of climate aid using UNFCCC classification: adaptation, climate change mitigation, cross-cutting, none.
  \item climate_aid_nok_mill: Estimated earmarked climate aid in million NOK (40 percent of significant objective).
  \item climate_adaptation_nok_mill: Estimated earmarked aid targeting adaptation (40 percent of significant objective).
  \item climate_mitigation_nok_mill: Estimated earmarked aid targeting mitigation (40 percent of significant objective).
  \item climate_mitigation_nok_mill_gross_fix: Handy for calculating the climate share of Norfund's investements. Estimated earmarked amounts extended targeting mitigation (negative extended amounts are excluded).
  \item climate_aid_nok_gross_fix: For reporting of climate finance the UNFCCC. Gross disbursements (negative extended amounts are excluded).
}
}
\description{
Add climate columns to dataframe
}
\examples{
?add_cols_climate()

}
