% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_climate_clean.R
\name{add_cols_climate_clean}
\alias{add_cols_climate_clean}
\title{Add climate columns to dataframe}
\usage{
add_cols_climate_clean(data)
}
\arguments{
\item{data}{Input dataframe of Norwegian development assistance}
}
\value{
Returns dataframe with additional climate-relevant columns:
\itemize{
  \item climate_aid_type: Type of climate aid using UNFCCC classification: adaptation, climate change mitigation, cross-cutting, none.
  \item climate_aid_nok_mill: Estimated earmarked climate aid in million NOK (40 percent of significant objective).
  \item climate_adaptation_nok_mill: Estimated earmarked aid targeting adaptation (40 percent of significant objective).
  \item climate_mitigation_nok_mill: Estimated earmarked aid targeting mitigation (40 percent of significant objective).
  \item climate_mitigation_nok_mill_gross: Handy for calculating the climate share of Norfund's investements. Estimated earmarked amounts extended targeting mitigation
  \item climate_adaptation_nok_mill_gross: Handy for calculating the climate share of Norfund's investements. Estimated earmarked amounts extended targeting adaptation
  \item climate_finance_nok_gross: For reporting of climate finance the UNFCCC. Gross disbursements
}
}
\description{
Add climate columns to dataframe
}
\examples{
?add_cols_climate_clean()

}
