% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_climate_finance.R
\name{add_cols_climate_finance}
\alias{add_cols_climate_finance}
\title{Add public climate finance columns to an existing statsys data frame}
\usage{
add_cols_climate_finance(df_statsys)
}
\arguments{
\item{df_statsys}{A statsys data frame, which must already be loaded into the environment.}
}
\value{
A statsys data frame with an additional `climate_finance_nok` column.
}
\description{
This function takes an existing statsys data frame as input and adds public climate finance columns using the UNFCCC methodology, including both Official Development Finance (ODA and OOF).
Earmarked climate finance is calculated using the Rio Markers for Climate Change adaptation and mitigation, applying a 40 percent coefficient for activities with only a significant climate change objective(s).
The Riomarked capitalisation(s) of Norfund Climate Investment Mandate/Climate Investment Fund is excluded (avoiding double counting), as the CIF investments are already included (OOF).
Imputed multilateral climate finance is calculated using the OECD imputed multilateral shares, along with Norad's temporary estimates for the most recent year(s).
All amounts are expressed in gross disbursements, and any negative gross disbursements in the statsys database are set to 0.
Note: Climate finance to all developing countries is included, including countries that are not non-Annex 1 parties (such as Ukraine), and should therefore be excluded when reporting to the UNFCCC.
The function returns the statsys data frame with the following additional columns for climate finance:
\describe{
  \item{\code{climate_finance_nok}}{Numeric variable of total public climate finance (earmarked and imputed multilateral).}
  \item{\code{climate_finance_tag}}{Logical variable to identify climate finance activities, mening Rio-marked activites (regardless of amounts extended) and positive climate shares.}
  \item{\code{climate_finance_channel}}{Categorical variable to separate earmarked from imputed multilateral climate finance.}
  \item{\code{climate_finance_channel2}}{Categorical variable to separate earmarked (ex. Norfund/KIF), Norfund/KIF, and imputed multilateral climate finance.}
  \item{\code{climate_finance_type_of_support_3levels}}{Categorical variable of type of support (UNFCCC levels): adaptation only, mitigation only, and cross-cutting. Only useful for \code{climate_finance_nok}, not for \code{climate_adaptation_finance_earmarked_nok} and \code{climate_mitigation_finance_earmarked_nok}.}
  \item{\code{climate_adaptation_finance_earmarked_nok}}{Numeric variable of earmarked climate adaptation finance. Note that many of the activities may also be cross-cutting, also aimed at climate change mitigation.}
  \item{\code{climate_mitigation_finance_earmarked_nok}}{Numeric variable of earmarked climate mitigation finance. Note that many of the activities may also be cross-cutting, also aimed at climate change adaptation.}
}
}
\details{
## Important:
Before running this function, you must have already loaded the statsys data frame by using
\code{noradstats::read_statsys()}.
}
\examples{
# Load the statsys data
df_statsys <- read_statsys()

# Add the climate_finance_nok column to the df_statsys data
df_statsys_climate_finance <- add_cols_climate_finance(df_statsys)

# Ensure that the df_statsys data frame is available before running this function,
# using the `read_statsys()` function
}
