% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_climate_oda.R
\name{add_cols_climate_oda}
\alias{add_cols_climate_oda}
\title{Add public climate oda columns to an existing statsys data frame}
\usage{
add_cols_climate_oda(df_oda)
}
\arguments{
\item{df_statsys}{An ODA data frame, which must already be loaded into the environment.}
}
\value{
An ODA data frame with an additional `climate_oda_nok` column.
}
\description{
This function takes an existing statsys data frame as input and adds public climate ODA columns.
Earmarked climate ODA is calculated using the Rio Markers for Climate Change adaptation and mitigation, applying a 40 percent coefficient for activities with only a significant climate change objective(s).
Imputed multilateral climate ODA is calculated using the OECD imputed multilateral shares, along with Norad's temporary estimates for the most recent year(s).
Imputed ordinary Norfund climate ODA (2 year averages) is calculated by Norad using the imputed multilateral shares methodology.
All amounts are expressed in net disbursements.
The function returns the statsys data frame with the following additional columns for climate ODA:
\describe{
  \item{\code{climate_oda_nok}}{Numeric variable of total public climate ODA (earmarked incl imputed ordinary Norfund and imputed multilateral).}
  \item{\code{climate_oda_tag}}{Logical variable to identify climate ODA activities, mening Rio-marked activites (regardless of disbursements) and positive imputed multilateral or imputed ordinary Norfund climate shares.}
  \item{\code{climate_oda_channel}}{Categorical variable to separate earmarked from imputed multilateral climate ODA}
  \item{\code{climate_oda_channel2}}{Categorical variable to separate earmarked (ex. capitalisation of ordinary Norfund), imputed ordinary Norfund ODA, and imputed multilateral ODA}
  \item{\code{climate_oda_type_of_support_3levels}}{Categorical variable of type of support (UNFCCC levels): adaptation only, mitigation only, and cross-cutting. Only useful for \code{climate_oda_nok}, not for \code{climate_adaptation_oda_earmarked_nok} and \code{climate_mitigation_oda_earmarked_nok}.}
  \item{\code{climate_oda_finance_earmarked_nok}}{Numeric variable of earmarked ODA for climate adaptation. Note that many of the activities may also be cross-cutting, also aimed at climate change mitigation.}
  \item{\code{climate_oda_finance_earmarked_nok}}{Numeric variable of earmarked ODA for climate mitigation, including imputed ordinary Norfund. Note that many of the activities may also be cross-cutting, also aimed at climate change adaptation.}
}
}
\details{
## Important:
Before running this function, you must have already loaded the ODA data frame by using
\code{noradstats::read_oda()}.
}
\examples{
# Load the ODA data
df_oda <- read_oda()

# Add the climate_oda_nok columns to the df_oda data frame
df_oda_climate_oda <- add_cols_climate_oda(df_oda)

# Ensure that the df_oda data frame is available before running this function,
# using the `read_oda()` function
}
