% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_conflict.R
\name{add_cols_conflict}
\alias{add_cols_conflict}
\title{Add UCDP conflict variables (country-year level) to ODA data frame from 1989 and later. From UCDP Georeferenced Event Dataset (GED)}
\usage{
add_cols_conflict(data, type = "statebased")
}
\arguments{
\item{data}{Input dataframe of Norwegian development assistance}

\item{type}{Type of violence: "statebased", "onesided", "nonstate", "all". Default is "statebased" conflict variables.}
}
\value{
Returns filtered data frame (Year >= 1989) with additional conflict columns based on the \emph{type} argument:
\itemize{
  \item *: violence/statebased_conflict/nonstate_conflict/onesided_violence: Logical variable. Minimum 1 statebased/nonstate/onsided conflict/violence ID in country-year.
  \item *_fatality_count: Number of fatalities in country-year. Numerical variable. Should not be used for summarisation in the agreement-level ODA data frame as the number is at country-level.
  \item *_count: Number of conflicts/organised violence IDs in country-year. Numerical variable. Should not be used for summarisation in the agreement-level ODA data frame as the number is at country-level.
  \item *_intensity: Intensity of the largest conflict in country-year. Categorical variable. Major/war (>1 000 fatalities) vs minor (<1 000 fatalities).
  \item *_intensity_sum: Intensity of the sum of conflicts in country-year. Categorical variable. Major/war (>1 000 fatalities) vs minor (<1 000 fatalities).
  \item statebased_intensity_lag2years: Intensity of the largest statebased conflict in country-year including a two year lag based on the statebased_intensity variable. Categorical variable. War (>1 000 fatalities) vs minor (<1 000 fatalities).
}
}
\description{
Add UCDP conflict variables (country-year level) to ODA data frame from 1989 and later. From UCDP Georeferenced Event Dataset (GED)
}
\examples{
?add_cols_conflict()
}
