% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_countrycode.R
\name{add_cols_countrycode}
\alias{add_cols_countrycode}
\title{Add country code columns to dataframe}
\usage{
add_cols_countrycode(data)
}
\arguments{
\item{data}{Input dataframe of Norwegian development assistance}
}
\value{
Returns dataframe with additional columns:
\itemize{
  \item ISO3: ISO-3 character code, identified using package \emph{countrycode} based on column \emph{Recipient country}. Non-matches are given NA values and are returned in a warning message. 
}
}
\description{
Add country code columns to dataframe
}
\examples{
?add_cols_countrycode()
}
