% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_education.R
\name{add_cols_education}
\alias{add_cols_education}
\title{Add column \emph{Education incl. Education in Emergencies} to dataframe}
\usage{
add_cols_education(data)
}
\arguments{
\item{data}{Input dataframe of Norwegian development assistance}
}
\value{
Returns dataframe with additional column:
\itemize{
  \item education_incl_emergencies: Includes \emph{Target area = Education} and \emph{DAC Sub sector (code+name) = 12 - Education in Emergencies}. 3 levels: 1) \emph{Education}, 2) \emph{Education in Emergencies}, 3) \emph{None}.
}
}
\description{
Add column \emph{Education incl. Education in Emergencies} to dataframe
}
\examples{
?add_cols_education()

}
