% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_food.R
\name{add_cols_food}
\alias{add_cols_food}
\title{Add food columns to an existing oda data frame}
\usage{
add_cols_food(df_oda)
}
\arguments{
\item{df_oda}{A oda data frame, which must already be loaded into the environment.}
}
\value{
A oda data frame with additional food columns.
}
\description{
This function takes an existing oda data frame as input and adds food columns.
The function returns the oda data frame with the following additional columns for food:
\describe{
  \item{\code{food_nok}}{Numeric variable of total disbursed food oda (earmarked and multilateral).}
  \item{\code{food_tag}}{Logical variable to identify food activities, meaning earmarked and multilateral (long-term and emergency)}
  \item{\code{food_oda_channel_highlevel}}{Categorical variable to separate earmarked oda to food from  multilateral oda to food.}
  \item{\code{food_oda_channel_lowlevel}}{Categorical variable to separate earmarked and multilateral and long-term and emergency.}
}
}
\details{
## Important:
Before running this function, you must have already loaded the oda data frame by using
\code{noradstats::read_oda()}.
}
\examples{
# Load the oda data
df_oda <- read_oda()

# Add food column to the df_oda data
df_oda_food <- add_cols_food(df_oda)

# Ensure that the df_oda data frame is available before running this function,
# using the `read_oda()` function
}
