% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_health.R
\name{add_cols_health}
\alias{add_cols_health}
\title{Add health columns to an existing oda data frame}
\usage{
add_cols_health(df_oda)
}
\arguments{
\item{df_oda}{A statsys data frame, which must already be loaded into the environment.}
}
\value{
A statsys data frame with additional health columns.
}
\description{
This function takes an existing oda data frame as input and adds health columns.
The function returns the oda data frame with the following additional columns for health:
\describe{
  \item{\code{health_nok}}{Numeric variable of total disbursed health oda (earmarked and imputed multilateral).}
  \item{\code{health_tag}}{Logical variable to identify health activities, meaning earmarked ODA for DAC sectors of long term and emergency aid and imputed multilateral ODA for health (positive imputed multilateral health shares)}
  \item{\code{health_oda_channel}}{Categorical variable to separate earmarked oda to health from imputed multilateral oda to health.}
  \item{\code{health_oda_channel2}}{Categorical variable to separate earmarked humanitarian oda to health, long term oda to health, and imputed multilateral to health.}
}
}
\details{
## Important:
Before running this function, you must have already loaded the statsys data frame by using
\code{noradstats::read_oda()}.
}
\examples{
# Load the statsys data
df_oda <- read_oda()

# Add health column to the df_oda data
df_oda_health <- add_cols_health(df_oda)

# Ensure that the df_oda data frame is available before running this function,
# using the `read_oda()` function
}
