% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_hum.R
\name{add_cols_hum}
\alias{add_cols_hum}
\title{Add humanitarian (hum) columns to an existing oda data frame}
\usage{
add_cols_hum(df_oda)
}
\arguments{
\item{df_oda}{A oda data frame, which must already be loaded into the environment.}
}
\value{
A oda data frame with additional hum columns.
}
\description{
This function takes an existing oda data frame as input and adds hum columns.
The function returns the oda data frame with the following additional columns for hum:
\describe{
  \item{\code{hum_nok}}{Numeric variable of total disbursed hum oda (earmarked and imputed multilateral).}
  \item{\code{hum_tag}}{Logical variable to identify hum activities, meaning earmarked ODA for DAC sectors and imputed multilateral ODA for hum (positive imputed multilateral hum shares)}
  \item{\code{hum_oda_channel}}{Categorical variable to separate earmarked oda to hum from imputed multilateral oda to hum.}
}
}
\details{
## Important:
Before running this function, you must have already loaded the oda data frame by using
\code{noradstats::read_oda()}.
}
\examples{
# Load the oda data
df_oda <- read_oda()

# Add hum columns to the df_oda data
df_oda_hum <- add_cols_hum(df_oda)

# Ensure that the df_oda data frame is available before running this function,
# using the `read_oda()` function
}
