% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_ldc.R
\name{add_cols_ldcs}
\alias{add_cols_ldcs}
\title{Add LDCs columns to an existing oda data frame}
\usage{
add_cols_ldcs(df_oda)
}
\arguments{
\item{df_oda}{A oda data frame, which must already be loaded into the environment.}
}
\value{
A oda data frame with additional LDCs columns.
}
\description{
This function takes an existing oda data frame as input and adds LDCs columns.
The function returns the oda data frame with the following additional columns for LDCs:
\describe{
  \item{\code{ldcs_nok}}{Numeric variable of total disbursed oda to LDCs (earmarked and imputed multilateral).}
  \item{\code{ldcs_tag}}{Logical variable to identify LDCs activities, meaning earmarked ODA for LDCs and imputed multilateral ODA for LDCs}
  \item{\code{ldcs_channel}}{Categorical variable to separate earmarked oda to LDCs from imputed multilateral oda to LDCs}
}
}
\details{
## Important:
Before running this function, you must have already loaded the oda data frame by using
\code{noradstats::read_oda()}.
}
\examples{
# Load the oda data
df_oda <- read_oda()

# Add LDCs column to the df_oda data
df_oda_ldcs <- add_cols_ldcs(df_oda)

# Ensure that the df_oda data frame is available before running this function,
# using the `read_oda()` function
}
