% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_srhr.R
\name{add_cols_srhr}
\alias{add_cols_srhr}
\title{Add srhr columns to an existing oda data frame}
\usage{
add_cols_srhr(df_oda)
}
\arguments{
\item{df_oda}{A oda data frame, which must already be loaded into the environment.}
}
\value{
A oda data frame with additional srhr columns.
}
\description{
This function takes an existing oda data frame as input and adds srhr columns.
The function returns the oda data frame with the following additional columns for srhr:
\describe{
  \item{\code{srhr_nok}}{Numeric variable of total disbursed srhr oda. Using hard-coded percentages for individual partners/sectors}
  \item{\code{srhr_tag}}{Logical variable to identify srhr activities, by identifying sectors, partners and agreements)}
  \item{\code{srhr_oda_channel_lowlevel}}{Categorical variable to separate the sectors and partners}
  \item{\code{srhr_oda_channel_highlevel}}{Categorical variable to separate earmarked ODA to SRHR and multilateral ODA to SRHR (estimate)}
}
}
\details{
## Important:
SRHR support to GFATM and GFF is hardcoded using agreement numbers. New agreements with these partners must be hardcoded in the noradstats function noradstats::add_cols_srhr() before using this function.
Before running this function, you must have already loaded the oda data frame by using
\code{noradstats::read_oda()}.
}
\examples{
# Load the oda data
df_oda <- read_oda()

# Add srhr column to the df_oda data
df_oda_health <- add_cols_srhr(df_oda)

# Ensure that the df_oda data frame is available before running this function,
# using the `read_oda()` function
}
