% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_norfund_climate_ratio_to_db.R
\name{calculate_climate_ratio}
\alias{calculate_climate_ratio}
\title{Calculate Climate Mitigation Ratio for Norfund DIM (internal function)}
\usage{
calculate_climate_ratio(df_dim)
}
\value{
A data frame containing the climate mitigation ratios.
}
\description{
This is an **internal helper function** and is not exported for direct use. It calculates
the two-year climate mitigation ratio for the Norfund DIM data.
}
