% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_norfund_dim_portfolio_climate_share_to_db.R
\name{calculate_climate_share}
\alias{calculate_climate_share}
\title{Calculate Climate Mitigation Share for Norfund DIM (internal function)}
\usage{
calculate_climate_share(df_norfund_dim)
}
\value{
A data frame containing the climate mitigation shares.
}
\description{
This is an **internal helper function** and is not exported for direct use. It calculates
the two-year climate mitigation share for the Norfund DIM data.
Removes rows with NA climate_share_2yr_avg (2014, the first year)
}
