% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crs_data_to_db.R
\name{create_crs_data_to_db}
\alias{create_crs_data_to_db}
\title{Function to save and overwrite CRS data to DuckDB Database}
\usage{
create_crs_data_to_db(df_crs)
}
\arguments{
\item{df_crs}{Name of the crs data frame in the R environment to be saved to the DuckDB database.}
}
\description{
Before running this function, load CRS data into a data frame in the R Environment:
1) Download CRS-parquet.parquet file from the OECD database available on CRS: Creditor Reporting System (flows) [cloud replica].
2) Read the CRS-parquet.parquet file into R Environment using nanoparquet::read_parquet()
}
\details{
The function checks if both the crs input data frame exists and then saves the data frame to
the DuckDB database table crs, overwriting the existing crs table.
}
\examples{
\dontrun{
# Call the function to save and overwrite the Statsys data
create_crs_data_to_db(df_crs)
}

}
