% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_imputed_multi_marker_to_db.R
\name{create_imputed_multi_marker_to_db}
\alias{create_imputed_multi_marker_to_db}
\title{Import and save imputed multilateral marker data into DuckDB database}
\usage{
create_imputed_multi_marker_to_db(
  filepath = "C:/Users/aaw262/Documents/R/norfund_climate/imputed_multi_marker.xlsx"
)
}
\arguments{
\item{filepath}{A string. The path to the Excel file containing the imputed multilateral marker data. 
The Excel file must contain the following columns:
- `aid_type`: Character string representing the type of aid.
- `agreement_partner`: Character string representing the partner involved in the agreement.
- `marker`: Character string representing the multilateral marker, meaning climate, environment etc.
- `year`: Integer representing the year.
- `share`: Numeric value representing the share of the marker.}
}
\value{
A data frame containing the imported multilateral marker data with the following columns:
`aid_type`, `agreement_partner`, `marker`, `year`, and `share`.
}
\description{
This is the exported main function that calls internal helper functions to import and save 
the imputed multilateral marker data into the DuckDB database.
}
\details{
Steps:
1. Imports the imputed multilateral marker data from an Excel spreadsheet. 
   The file must contain the columns: `aid_type`, `agreement_partner`, `marker`, `year`, and `share`.
   The imported spreadsheet should not include any observations with NA share, only real values.
   The share for each agreement partner is calculated by the OECD secretariate and available online.
2. Saves the imported data into the DuckDB database under the table name 'imputed_multi_marker_ratio'.


This function imports an Excel spreadsheet containing imputed multilateral marker data, 
which must have the following columns: `aid_type`, `agreement_partner`, `marker`, `year`, and `share`.

Once imported, the data is saved into a table named `imputed_multi_marker_ratio` in the DuckDB database. 
This table will overwrite any existing data with the same name in the database.
}
\examples{
\dontrun{
# Example: Import and save multilateral marker data from a file
create_imputed_multi_marker_to_db("path/to/imputed_multi_marker.xlsx")
}

}
