% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_norfund_dim_portfolio_climate_ratio_to_db.R
\name{create_norfund_dim_portfolio_climate_ratio_to_db}
\alias{create_norfund_dim_portfolio_climate_ratio_to_db}
\title{Create a table of the Norfund DIM Portfolio Climate Ratio and store it in the DuckDB database}
\usage{
create_norfund_dim_portfolio_climate_ratio_to_db(
  cim_filepath = "agreement_number_norfund_cim.xlsx"
)
}
\arguments{
\item{cim_filepath}{A string. The path to the Excel file containing the CIM agreements to exclude. Default is "avtalenr_cim.xlsx". The file must contain the two columns *agreement_number* and *cim_dim*.}
}
\value{
This function does not return a value. It saves the climate ratio table to the DuckDB database 
and displays a success message upon completion.
}
\description{
This function calculates the annual climate mitigation ratio (2 year averages) of the Norfund DIM (development mandate) portfolio and saves the year-ratio data frame in the DuckDB database.
The user must provide an Excel file containing the Norfund CIM (Climate Investment Mandate) agreements to be excluded.
}
\details{
After running this function, the 'norfund_dim_climate_ratio' table will be created or overwritten in 
the DuckDB database, and a success message will be displayed.
The `create_norfund_dim_portfolio_climate_ratio_to_db()` function performs the following steps using internal helper functions:
\itemize{
  \item `import_cim_data()`: Imports the CIM agreements data from an Excel file.
  \item `prepare_norfund_dim_data()`: Prepares the Norfund DIM data by importing statsys data, filtering out CIM agreements from 2022 onwards, 
  and performing necessary data processing.
  \item `calculate_climate_ratio()`: Calculates the two-year climate mitigation ratio.
  \item `save_climate_ratio_to_db()`: Saves the calculated climate ratio data to the DuckDB database.
}
}
\examples{
\dontrun{
# Use default CIM file path
create_norfund_dim_portfolio_climate_ratio_to_db()

# Specify a different CIM file path
create_norfund_dim_portfolio_climate_ratio_to_db("path/to/cim_file.xlsx")
}
}
