% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pta_agreement_totals_to_db.R
\name{create_pta_agreement_totals_to_db}
\alias{create_pta_agreement_totals_to_db}
\title{Function to import, save and overwrite PTA Agreement totals data to DuckDB Database}
\usage{
create_pta_agreement_totals_to_db(input_csv)
}
\arguments{
\item{input_csv}{A string. Path to the input CSV file containing the Statsys data.}
}
\description{
Reads PTA Agreement totals from a CSV file, processes it, cleans the column names, and then saves the data to a DuckDB database.
The existing data in the "pta_agreement_totals" table in the database will be overwritten.
}
\details{
Before running this function, it is recommended that users first inspect the
data by running the `read_pta_agreement_totals_from_csv()` function to ensure that the data
has the expected structure and column types.


The function checks if both the CSV file and the DuckDB database file exist. If either
file is not accessible, the function will stop and return an error message. It uses 
internal `noradstats` functions to read and process the data. 
The `janitor` package is used to clean column names. The resulting 
data is then written to the DuckDB database, overwriting the existing "pta_agreement_totals" table.
}
\examples{
\dontrun{
# Example usage:
# Path to the CSV file containing the Statsys data
input_csv <- "path/to/your_pta_report.csv"

# Call the function to save and overwrite the Statsys data
create_pta_agreement_totals_to_db(input_csv)
}

}
