% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_aiddata.R
\name{find_aiddata}
\alias{find_aiddata}
\title{Find available files on (noradstats) Google Drive}
\usage{
find_aiddata(gmail_account = "noradstats@gmail.com")
}
\arguments{
\item{gmail_account}{Google drive account. Default: \emph{noradstats@gmail.com}}
}
\value{
Returns a tibble of available files on Noradstats Google Drive
}
\description{
Find available files on (noradstats) Google Drive
}
\examples{
?find_aiddata()

}
