% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aiddata.R
\name{get_aiddata}
\alias{get_aiddata}
\title{Download data from noradstats google drive}
\usage{
get_aiddata(
  file = "statsys_ten.csv",
  path = NULL,
  gmail_account = "noradstats@gmail.com"
)
}
\arguments{
\item{file}{Character string of file to download. Default file is \emph{statsys_ten.csv}. Use noradstats::find_aiddata() to find available files.}

\item{path}{Character string of path to save file. Default path is file name in current directory.}

\item{gmail_account}{Google Drive account. Default is \emph{noradstats@gmail.com}}
}
\value{
Downloads file to path.
}
\description{
Download data from noradstats google drive
}
\examples{
?get_aiddata()

}
