% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_donors.R
\name{get_donors}
\alias{get_donors}
\title{Get data on OECD DAC donors ODA to countries and regions (DAC2a), using the OECD SDMX API. Amounts in USD million}
\usage{
get_donors(startyear = 2020, endyear = 2020)
}
\arguments{
\item{startyear}{Specify a numeric value of the first year in time period. Default value is \emph{2020}.}

\item{endyear}{Specity a numeric value of the last year in time period. Default value is \emph{2020}.}
}
\value{
Returns a dataframe (tibble) on OECD DAC donors ODA to countries and regions (DAC2a)
}
\description{
Get data on OECD DAC donors ODA to countries and regions (DAC2a), using the OECD SDMX API. Amounts in USD million
}
\examples{
?df_donors <- get_donors()

}
