% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_norfund_dim_portfolio_climate_share_to_db.R
\name{import_cim_data}
\alias{import_cim_data}
\title{Import CIM Data from Excel (internal function)}
\usage{
import_cim_data(cim_filepath)
}
\arguments{
\item{cim_filepath}{A string. The path to the Excel file.}
}
\value{
A data frame containing the CIM agreements.
}
\description{
This is an **internal helper function** and is not exported for direct use. It reads a table of 
Climate Investment Mandate (CIM) data from an Excel file.
}
