% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_imputed_norfund_climate_share_to_db.R
\name{import_norfund_capitalisation}
\alias{import_norfund_capitalisation}
\title{Import Norfund Capitalisation Data (Internal Function)}
\usage{
import_norfund_capitalisation(cap_dim_filepath)
}
\arguments{
\item{cap_dim_filepath}{A string. The path to the Excel file of agreement numbers of capitalisation of ordinary Norfund DIM.}
}
\value{
A data frame containing the capitalisation agreements of ordinary Norfund DIM.
}
\description{
This internal function imports capitalisation agreements data for Norfund DIM 
from an Excel file.
}
