% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noradstats_download.R
\name{noradstats_download}
\alias{noradstats_download}
\title{Download file from Google drive (to download development aid dataset)}
\usage{
noradstats_download(file = NULL, gmail_account = "noradstats@gmail.com")
}
\arguments{
\item{file}{Filename to download. Use noradstats::find to find available files.}

\item{gmail_account}{Google drive account. Default: noradstats@gmail.com}
}
\value{
Downloads the file and saves the file in working directory.
}
\description{
Download file from Google drive (to download development aid dataset)
}
\examples{
?noradstats_download()

}
