% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noradstats_find.R
\name{noradstats_find}
\alias{noradstats_find}
\title{Find files on Google drive (to find development aid datasets)}
\usage{
noradstats_find(gmail_account = "noradstats@gmail.com")
}
\arguments{
\item{gmail_account}{Google drive account. Default: noradstats@gmail.com}
}
\value{
Returns a tibble of available files on google drive account. Look for ODA datasets.
}
\description{
Find files on Google drive (to find development aid datasets)
}
\examples{
?noradstats_find()

}
