% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noradstats_read_csv.R
\name{noradstats_read_csv}
\alias{noradstats_read_csv}
\title{Read a delimited development aid dataset (csv) into a tibble}
\usage{
noradstats_read_csv(file)
}
\arguments{
\item{file}{Path to a local csv file}
}
\value{
Returns a tibble using the vroom package, using delim = ";", comma as decimal, and num_threads = 1.
}
\description{
Read a delimited development aid dataset (csv) into a tibble
}
\examples{
?noradstats_read_csv()

}
