% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_officialoda.R
\name{read_aidresults}
\alias{read_aidresults}
\title{Download and read the Official ODA statistics from aidresults.no into R}
\usage{
read_aidresults(download_folder, url = tenyears)
}
\arguments{
\item{download_folder}{Path to the default download folder in the default web browser (path/to/downloadfolder), typically C:/Users/xxx/Downloads.".}

\item{url}{The URL of the API call for the selected data. Find the URL for the selected data at https://resultater.norad.no/microdata. The default call is all data for the last ten years.}
}
\value{
Returns a tibble data frame of official ODA statistics from aidresults.no
}
\description{
The function opens the URL of the API call for the default or specified data in the default web browser and downloads the csv file to the default downloads folder in the browser.
The function requires the default download folder in the default web browser to be able to read the csv file into R.
The function will look for the new csv file in the default download folder and read the csv file into the R environment. It will stop after 60 attempts.
}
\examples{
?read_officialoda(download_folder = "path/to/downloadfolder", url = tenyears)

}
