% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aidresults_from_csv.R
\name{read_aidresults_from_csv}
\alias{read_aidresults_from_csv}
\title{Read aidresults.no data into R from user-specified CSV file}
\usage{
read_aidresults_from_csv(path)
}
\arguments{
\item{path}{Required path to CSV file of Offical Norwegian ODA data downloaded from aidresults.no: https://resultater.norad.no/microdata}
}
\value{
Returns a tibble of Norwegian ODA data
}
\description{
This function reads a CSV file of ODA statistics downloaded from aidresults.no: https://resultater.norad.no/microdata
The function excepcts the CSV to be delimited by semicolons (;) commas (,) as decimal marks and UTF-16LE encoding.
It expects the CSV file to be at the specified path and returns a tibble containing the data. Make sure the path is correctly specified to avoid errors.
The function checks for file existence and valid path input before attempting to read the file.
}
\examples{
?read_aidresults_from_csv
}
