% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_crs_database.R
\name{read_crs_database}
\alias{read_crs_database}
\title{Connecting to CRS database file and create a remote tibble linked to the crs_ten table in the database: International activity level CRS data (2013-2022)}
\usage{
read_crs_database()
}
\value{
Returns a remote table (tibble) linked to the 'crs_ten' table in the DuckDB database. Remember to use dbplyr syntax to collect data from the remote tibble and to disconnect when done, using DBI::dbDisconnect(con, shutdown=TRUE)
}
\description{
Connecting to CRS database file and create a remote tibble linked to the crs_ten table in the database: International activity level CRS data (2013-2022)
}
\examples{
?read_crs_database()
}
