% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_donors.R
\name{read_donors}
\alias{read_donors}
\title{Read international ODA data from OECD/DAC donors to countries and regions into R}
\usage{
read_donors(startyear = 2020, endyear = 2020)
}
\arguments{
\item{startyear}{Specify a numeric value of the first year in time period. Default value is \emph{2020}.}

\item{endyear}{Specity a numeric value of the last year in time period. Default value is \emph{2020}.}
}
\value{
Returns a dataframe (tibble) on OECD DAC donors ODA to countries and regions (DAC2a)
}
\description{
This function connects to the OECD SDMX API and extracts data DAC-members' ODA to countries and regions by year (DAC2a), including metadata. Amounts in USD million
}
\examples{
?df_donors <- read_donors()

}
