% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_imputed_countries.R
\name{read_imputed_countries}
\alias{read_imputed_countries}
\title{Read Norwegian imputed multilateral ODA to countries and regions data into R}
\usage{
read_imputed_countries(startyear = 2011, endyear = 2020)
}
\arguments{
\item{startyear}{Specify a numeric value of the first year in time period. Default value is \emph{2011}.}

\item{endyear}{Specity a numeric value of the last year in time period. Default value is \emph{2020}.}
}
\value{
Returns a dataframe (tibble) of Norwegian imputed multilateral ODA to countries and regions
}
\description{
This function connects to the OECD SDMX API and extracts Norwegian imputed multilateral ODA to countries and regions data by year, including metadata.
The amounts are USD million and caclulates NOK million using the exchange rate data from the OECD API.
}
\examples{
?df_imputed_countries <- read_imputed_countries()

}
