% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_imputed_multi_org_incomegroup_shares.R
\name{read_imputed_multi_org_incomegroup_shares}
\alias{read_imputed_multi_org_incomegroup_shares}
\title{Read Imputed Multilateral Income Group Shares by oganisation data into R}
\usage{
read_imputed_multi_org_incomegroup_shares()
}
\value{
Returns a tibble with two columns: `aid_type`, `agreement_partner`, `income_group`, `year` and `share`, from the 'imputed_multi_org_incomegroup_shares' table in the DuckDB database.
}
\description{
This function imports from the DuckDB database a data frame of the annual imputed multilateral org income group shares
calculated by the OECD-secretariate (except for Norad estimates for the last year). A column named income_gropu
identifies the income_group share, for instance: LDCs
}
\examples{
\dontrun{
# Read the imputed_multi_org_incomegroup_shares table from the DuckDB database:
df_imputed_multi_org_incomegroup_shares <- read_imputed_multi_org_incomegroup_shares()

# Display the first few rows of the data:
head(df_imputed_multi_org_incomegroup_shares)
}
}
