% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_imputed_multi_sector_shares.R
\name{read_imputed_multi_sector_shares}
\alias{read_imputed_multi_sector_shares}
\title{Read Imputed Multilateral Sector Shares data into R}
\usage{
read_imputed_multi_sector_shares()
}
\value{
Returns a tibble with two columns: `aid_type`, `agreement_partner`, `sector`, `year` and `share`, from the 'imputed_multi_sector_shares' table in the DuckDB database.
}
\description{
This function imports from the DuckDB database a data frame of the annual imputed multilateral sector shares
calculated by the OECD-secretariate (except for Norad estimates for the last year). A column named sector
identifies the sector share, for instance: health
}
\examples{
\dontrun{
# Read the imputed_multi_sector_shares table from the DuckDB database:
df_imputed_multi_sector_shares <- read_imputed_multi_sector_shares()

# Display the first few rows of the data:
head(df_imputed_multi_sector_shares)
}
}
