% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_imputed_multi_shares.R
\name{read_imputed_multi_shares}
\alias{read_imputed_multi_shares}
\title{Read Imputed Multilateral Shares data into R}
\usage{
read_imputed_multi_shares()
}
\value{
Returns a tibble with two columns: `year` and `climate_ratio`, from the 'imputed_multi_shares' table in the DuckDB database.
}
\description{
This function imports from the DuckDB database a data frame of the annual imputed multilateral shares
calculated by the OECD-secretariate (except for Norad estimates for the last year). A column named Marker
identifies the Marker share, for instance: climate.
}
\examples{
\dontrun{
# Read the imputed_multi_shares table from the DuckDB database:
df_imputed_multi_shares <- read_imputed_multi_shares()

# Display the first few rows of the data:
head(df_imputed_multi_shares)
}
}
