% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_imputed_sectors.R
\name{read_imputed_sectors}
\alias{read_imputed_sectors}
\title{Read Norwegian imputed multilateral ODA to sectors data into R}
\usage{
read_imputed_sectors(startyear = 2020, endyear = 2020)
}
\arguments{
\item{startyear}{Specify a numeric value of the first year in time period. Default value is \emph{2011}.}

\item{endyear}{Specity a numeric value of the last year in time period. Default value is \emph{2020}.}
}
\value{
Returns a tibble dataframe of Norwegian imputed multilateral ODA to sectors
}
\description{
This function connects to the OECD SDMX API and extracts Norwegian imputed multilateral ODA to sectors data by year, including metadata.
The amounts are USD million and caclulates NOK million using the exchange rate data from the OECD API.
}
\examples{
?df_imputed_sectors <- read_imputed_sectors()

}
