% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_international_crs.R
\name{read_international_crs}
\alias{read_international_crs}
\title{Read international CRS data into R}
\usage{
read_international_crs()
}
\value{
A tibble containing international CRS data.
}
\description{
This function connects to a specified DuckDB database table to retrieve CRS (Creditor Reporting System) activity level data of 
Official Development Finance (ODA/OOF) from official donors (DAC, non-DAC, multilaterals) and private philantropy, starting from 1960 to the recent year. 
It returns all the data in the database table as a tibble. The database file is expected to be located in a synchronized SharePoint directory on the user's local computer.
Data source: https://stats.oecd.org/DownloadFiles.aspx?DatasetCode=CRS1
}
\examples{
?read_oda()

}
