% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_norfund_climate_ratio.R
\name{read_norfund_climate_ratio}
\alias{read_norfund_climate_ratio}
\title{Read Norfund DIM Climate Ratio data into R}
\usage{
read_norfund_climate_ratio()
}
\value{
Returns a tibble with two columns: `year` and `climate_ratio`, from the 'norfund_dim_climate_ratio' table in the DuckDB database.
}
\description{
This function imports a data frame of the annual climate mitigation ratio (2-year averages) of the Norfund DIM 
(Development Investment Mandate) portfolio from the DuckDB database.
}
\examples{
\dontrun{
# Read the Norfund DIM Climate Ratio from the DuckDB database:
df_climate_ratio <- read_norfund_climate_ratio()

# Display the first few rows of the data:
head(df_climate_ratio)
}
}
