% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oda.R
\name{read_oda}
\alias{read_oda}
\title{Read Norwegian ODA data into R}
\usage{
read_oda()
}
\value{
Returns a tibble of Norwegian ODA data from 1960 to the recent year. Frame agreement level data is excluded.
}
\description{
This function imports all Norwegian Official Development (ODA) from the Statsys table in the DuckDB database.
This include ODA data from 1960 to the recent year. Frame agreement level data is excluded.
The DuckDB database file is located on Norads Microsoft Sharepoint site and is expected to be synced via Microsoft Teams to to the users local directory.
}
\examples{
?read_oda()

}
