% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/official_colnames.R
\name{read_official_colname_map}
\alias{read_official_colname_map}
\title{Read official column name mapping from package extdata}
\usage{
read_official_colname_map()
}
\description{
Read official column name mapping from package extdata
}
\keyword{internal}
