% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_officialoda.R
\name{read_officialoda}
\alias{read_officialoda}
\title{Read Official ODA statistics from aidresults.no to a tibble}
\usage{
read_officialoda(download_folder = NULL, url = tenyears)
}
\arguments{
\item{download_folder}{The path to the default download folder in the default web browser. No default path provided.}

\item{url}{The URL of the API call for the selected data. Find the URL for the selected data at https://resultater.norad.no/microdata. The default call is all data for the last ten years.}
}
\value{
The default web browser opens and generates a csv file containg the selected data and saves the file in the default download folder. The file is read as a tibble into the R environment.
}
\description{
Read Official ODA statistics from aidresults.no to a tibble
}
\examples{
?read_officialoda(download_folder = "path/to/downloadfolder", url = tenyears)

}
