% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pta_agreement_totals.R
\name{read_pta_agreement_totals}
\alias{read_pta_agreement_totals}
\title{Read PTA Agreement Totals data table from DuckDB into R}
\usage{
read_pta_agreement_totals()
}
\value{
Returns a tibble with 7 columns: `agreement_no`, `agreement_title`, `agreement_partner`, `agreement_period`, `agreement_period_from`, `agreement_period_to` and `expected_agreement_total`.
}
\description{
This function imports from the DuckDB database a data frame of PTA Agreement Totals from 1960 to today.
}
\examples{
\dontrun{
# Read the pta_agreement_totals table from the DuckDB database:
df_pta_agreement_totals <- read_pta_agreement_totals()

# Display the first few rows of the data:
head(df_pta_agreement_totals)
}
}
