% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pta_agreement_totals_from_csv.R
\name{read_pta_agreement_totals_from_csv}
\alias{read_pta_agreement_totals_from_csv}
\title{Read PTA Agreement totals report into R from user-specified CSV file}
\usage{
read_pta_agreement_totals_from_csv(path)
}
\arguments{
\item{path}{Path to the PTA Agreement totals CSV file}
}
\value{
A cleaned data frame with one row per agreement
}
\description{
Reads and processes the PTA Agreement totals report to produce a cleaned
dataset of agreement-level data.
}
