% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pta_disbursement_level.R
\name{read_pta_disbursement_level}
\alias{read_pta_disbursement_level}
\title{Read PTA Disbursement level data table from DuckDB into R}
\usage{
read_pta_disbursement_level()
}
\value{
Returns a tibble with many columns.
}
\description{
This function imports from the DuckDB database a data frame of PTA Disbursement level from 2000 and beyond
}
\examples{
\dontrun{
# Read the pta_disbursement_level table from the DuckDB database:
df_pta_disbursement_level <- read_pta_disbursement_level()

# Display the first few rows of the data:
head(df_pta_disbursement_level)
}
}
