% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pta_disbursement_level_from_csv.R
\name{read_pta_disbursement_level_from_csv}
\alias{read_pta_disbursement_level_from_csv}
\title{Read PTA Disbursement level report into R from user-specified CSV file}
\usage{
read_pta_disbursement_level_from_csv(path)
}
\arguments{
\item{path}{Path to the PTA Disbursement level CSV file}
}
\value{
A cleaned data frame with one row per agreement
}
\description{
Reads and processes the PTA Disbursement level report to produce a cleaned
dataset of disbursement level data
}
