% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statsys.R
\name{read_statsys}
\alias{read_statsys}
\title{Read Statsys data into R}
\usage{
read_statsys()
}
\value{
A tibble containing Norwegian ODA data.
}
\description{
This function connects to a specified DuckDB database table to retrieve Statsys data. This includes official development assistance (ODA), frame agreement level data, Other official flows(OOF), export credits and private flows.
The data is available from 1960 to the recent year.
The database file is expected to be located in a synchronized SharePoint directory on the user's local computer.
The function returns all the data in the database table as a tibble data frame.
}
\examples{
?read_statsys()

}
