% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statsys_from_csv.R
\name{read_statsys_from_csv}
\alias{read_statsys_from_csv}
\title{Read Statsys data into R from user-specified CSV file}
\usage{
read_statsys_from_csv(path)
}
\arguments{
\item{path}{Required path to CSV file of Statsys data.}
}
\value{
Returns a tibble of Statsys data.
}
\description{
This function reads a Statsys CSV file, which should be delimited by semicolons (;)
and have commas (,) as decimal marks. It expects the CSV file to be at the specified path
and returns a tibble containing the data. Make sure the path is correctly specified to avoid errors.
The function checks for file existence and valid path input before attempting to read the file.
}
\examples{
statsys_data <- read_statsys_from_csv("path/to/your/statsys_file.csv")

}
