% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_norfund_dim_portfolio_climate_ratio_to_db.R
\name{save_climate_ratio_to_db}
\alias{save_climate_ratio_to_db}
\title{Save Climate Mitigation Ratio Data to DuckDB (internal function)}
\usage{
save_climate_ratio_to_db(df_climate_ratio)
}
\description{
This is an **internal helper function** and is not exported for direct use. It saves
the calculated Norfund DIM climate ratio data to the DuckDB database and diconnects.
}
