% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_international_crs.R
\name{access_international_crs}
\alias{access_international_crs}
\title{Access database of international CRS data from R}
\usage{
access_international_crs()
}
\value{
Returns a proxy tibble connected to the CRS table in the DuckDB database.
}
\description{
This function creates a proxy tibble connected to the CRS table in the DuckDB database.
This data includes CRS (Creditor Reporting System) activity level data of Official Development Finance (ODA/OOF) from official donors (DAC, non-DAC, multilaterals) and private philantropy, starting from 1973 to the recent year.
The DuckDB database file is located on Norads Microsoft Sharepoint site and is expected to be synced via Microsoft Teams to to the users local directory.
Use DBI::dbDisconnect(con, shutdown=TRUE) to close connection to database.
Data source: https://stats.oecd.org/DownloadFiles.aspx?DatasetCode=CRS1
}
\examples{
?access_crs()

}
