% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_education.R
\name{add_cols_education}
\alias{add_cols_education}
\title{Add education columns to an existing oda data frame}
\usage{
add_cols_education(df_oda)
}
\arguments{
\item{df_oda}{A statsys data frame, which must already be loaded into the environment.}
}
\value{
A statsys data frame with additional education columns.
}
\description{
This function takes an existing oda data frame as input and adds education columns.
The function returns the oda data frame with the following additional columns for education:
\describe{
  \item{\code{education_nok}}{Numeric variable of total disbursed education oda (earmarked and imputed multilateral).}
  \item{\code{education_tag}}{Logical variable to identify education activities, meaning earmarked ODA for DAC sectors of long term and emergency aid and imputed multilateral ODA for education (positive imputed multilateral education shares)}
  \item{\code{education_oda_channel}}{Categorical variable to separate earmarked oda to education from imputed multilateral oda to education.}
  \item{\code{education_oda_channel2}}{Categorical variable to separate earmarked humanitarian oda to education, long term oda to education, and imputed multilateral to education.}
}
}
\details{
## Important:
Before running this function, you must have already loaded the statsys data frame by using
\code{noradstats::read_oda()}.
}
\examples{
# Load the statsys data
df_oda <- read_oda()

# Add education column to the df_oda data
df_oda_education <- add_cols_education(df_oda)

# Ensure that the df_oda data frame is available before running this function,
# using the `read_oda()` function
}
