% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_violence.R
\name{add_cols_violence}
\alias{add_cols_violence}
\title{Add UCDP conflict variables (country-year level) to ODA data frame from 1989 and later. From UCDP Georeferenced Event Dataset (GED)}
\usage{
add_cols_violence(data)
}
\arguments{
\item{data}{Input dataframe of Norwegian development assistance}
}
\value{
Returns filtered data frame (Year >= 1989) with additional conflict columns:
\itemize{
  \item violence: Logical Variable where all observations are TRUE to identify the country-years with at least 1 conflict ID of statebased/nonstate/onesided violence.
  \item violence_fatality_count_max: Numerical variable of number of fatalities from the larges conflict in country-year. Should not be used for summarisation in the agreement-level ODA data frame as the number is at country-level.
  \item violence_fatality_count_sum: Numerical variable of number of fatalities from all conflicts in country-year. Should not be used for summarisation in the agreement-level ODA data frame as the number is at country-level.
  \item violence_intensity_max: Categorical variable of intensity in fatalities of the largest conflict in country-year. "war" >= 1000, "major" > 150 < 1000, "minor" < 150. Non-conflicts are NA. The 150 fatality treshold is based on the World bank methodology: https://thedocs.worldbank.org/en/doc/fb0f93e8e3375803bce211ab1218ef2a-0090082023/original/Classification-of-Fragility-and-Conflict-Situations-FY24.pdf
  \item violence_intensity_sum: Categorical variable of intensity in fatalities from all conflicts in country-year. "war" >= 1000, "major" > 150 < 1000, "minor" < 150. Non-conflicts are NA. The 150 fatality treshold is based on the World bank methodology: https://thedocs.worldbank.org/en/doc/fb0f93e8e3375803bce211ab1218ef2a-0090082023/original/Classification-of-Fragility-and-Conflict-Situations-FY24.pdf
}
}
\description{
Add UCDP conflict variables (country-year level) to ODA data frame from 1989 and later. From UCDP Georeferenced Event Dataset (GED)
}
\examples{
?add_cols_violence()
}
