% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aiddata.R
\name{download_aiddata}
\alias{download_aiddata}
\title{Download file from (noradstats) Google Drive}
\usage{
download_aiddata(
  file = "statsys_10yr.csv",
  gmail_account = "noradstats@gmail.com"
)
}
\arguments{
\item{file}{Character string of desired file to download. Default: \emph{statsys_10yr.csv}. Use noradstats::find_file() to find available files.}

\item{gmail_account}{Google Drive account. Default: noradstats@gmail.com}
}
\value{
Downloads and saves file in working directory.
}
\description{
Download file from (noradstats) Google Drive
}
\examples{
?download_aiddata()

}
