% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_aiddata.R
\name{find_aiddata}
\alias{find_aiddata}
\title{Find files on noradstats google drive}
\usage{
find_aiddata(gmail_account = "noradstats@gmail.com")
}
\arguments{
\item{gmail_account}{Google drive account. Default is \emph{noradstats@gmail.com}}
}
\value{
Returns a tibble of files on noradstats google drive
}
\description{
Find files on noradstats google drive
}
\examples{
?find_aiddata()

}
