% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duckdb_path.R
\name{get_duckdb_path}
\alias{get_duckdb_path}
\title{Get the user-specific path to the DuckDB database}
\usage{
get_duckdb_path()
}
\value{
A string containing the full path to the DuckDB database.
}
\description{
This function determines the user-specific file path to the DuckDB database 
based on the operating system and user home directory.
}
