% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_norfund_dim_portfolio_climate_ratio_to_db.R
\name{import_cim_data}
\alias{import_cim_data}
\title{Import CIM Data from Excel (internal function)}
\usage{
import_cim_data(filepath = "agreement_number_norfund_cim.xlsx")
}
\value{
A data frame containing the CIM agreements.
}
\description{
This is an **internal helper function** and is not exported for direct use. It reads a table of
Climate Investment Mandate (CIM) data from an Excel file.
}
\details{
The Excel file must contain two columns: `agreement_number` and `cim_dim`.
}
