% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_norfund_dim_portfolio_climate_share_to_db.R
\name{prepare_norfund_dim_data}
\alias{prepare_norfund_dim_data}
\title{Prepare Norfund DIM Data (internal function)}
\usage{
prepare_norfund_dim_data(df_cim)
}
\value{
A data frame of the Norfund DIM data.
}
\description{
This is an **internal helper function** and is not exported for direct use. It imports statsys data and
process the data to include only relevant Norfund DIM data (Development Investment Mandate (DIM).
Excludes CIF agreements from 2022 and onwards (keeping old agreements before CIM was created - as there are old agreements also transfered from DIM to CIM in 2022)
}
