% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aiddata.R
\name{read_aiddata}
\alias{read_aiddata}
\title{Read Norwegian Aid data (Statsys csv) to a tibble}
\usage{
read_aiddata(file = path)
}
\arguments{
\item{file}{csv file of Statsys data. Default is \emph{statsys_ten.csv}}
}
\value{
Returns a tibble using the readr::read_delim function. Using delim = ";", specifying type of specific columns and decimal mark.
}
\description{
Read Norwegian Aid data (Statsys csv) to a tibble
}
\examples{
?read_aiddata()

}
