% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aiddata.R
\name{read_aiddata}
\alias{read_aiddata}
\title{Read delimited Norwegian Development Aid dataset (Statsys format and column names) into a tibble}
\usage{
read_aiddata(file = "statsys_10yr.csv")
}
\arguments{
\item{file}{Path to local csv-file of Norwegian Development Aid (Statsys format and column names). Default: \emph{statsys_10yr.csv}}
}
\value{
Returns a tibble using the vroom package, using delim = ";", comma as decimal, and num_threads = 1.
}
\description{
Read delimited Norwegian Development Aid dataset (Statsys format and column names) into a tibble
}
\examples{
?read_aiddata()

}
