% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_crs_database.R
\name{read_crs_database}
\alias{read_crs_database}
\title{Read international activity level CRS data (1973-). Create a remote tibble connected to the crs table in the DuckDB database file stored in Sharepoint}
\usage{
read_crs_database()
}
\value{
Returns a remote table (tibble) linked to the 'crs' table in the DuckDB database. Use dplyr syntax to collect data from the remote tibble and to disconnect when done, using DBI::dbDisconnect(con, shutdown=TRUE)
}
\description{
Read international activity level CRS data (1973-). Create a remote tibble connected to the crs table in the DuckDB database file stored in Sharepoint
}
\examples{
?read_crs_database()
}
