% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oda.R
\name{read_oda}
\alias{read_oda}
\title{Read Norwegian ODA data into R}
\usage{
read_oda()
}
\value{
A tibble containing Norwegian ODA data.
}
\description{
This function connects to a specified DuckDB database table to retrieve Norwegian Official Development
Assistance (ODA) activity level data, starting from 1960 to the recent year. It returns all the data in the database table as a tibble.
The database file is expected to be located in a synchronized SharePoint directory on the user's local computer.
}
\examples{
?read_oda()

}
