% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statsys.R
\name{read_statsys}
\alias{read_statsys}
\title{Read Statsys data into R}
\usage{
read_statsys()
}
\value{
Returns a tibble of ODA, OOF and PF data from the statsys table in the DuckDB database.
}
\description{
This function imports all data from the statsys table in the DuckDB database.
This data includes Norwegian official development assistance (ODA), ODA frame agreement level data, Other official flows(OOF), export credits and private flows.
The data covers 1960 to recent year.
The DuckDB database file is located on Norads Microsoft Sharepoint site and is expected to be synced via Microsoft Teams to to the users local directory.
}
\examples{
?read_statsys()

}
